==> Note Books/8 Data Preprocessing.ipynb <==
# Load processed data from previous step
import pandas as pd
Data_path = "/home/danial/Data Science/Churn Prediction/Data/Processed/telco_featured.csv"
df = pd.read_csv(Data_path)
df.head()


# Identify categorical columns and apply One-Hot Encoding
categorical_cols = df.select_dtypes(include=['object']).columns
df_encoded = pd.get_dummies(df, columns=categorical_cols, drop_first=True)
df_encoded.head()


# Scale numerical features using StandardScaler
from sklearn.preprocessing import StandardScaler

num_cols = df_encoded.select_dtypes(include=['int64','float64']).columns
scaler = StandardScaler()
df_encoded[num_cols] = scaler.fit_transform(df_encoded[num_cols])
df_encoded.head()


num_cols = ['tenure','MonthlyCharges','TotalCharges','AvgChargesPerMonth','ChargeDiff','ChargeDiffPerc','NumServices','HasInternet','ChurnFlag','gender_Male']
corr_matrix = df_encoded[num_cols].corr().abs()
corr_matrix


# Drop highly correlated or duplicate columns
cols_to_drop = ['AvgChargesPerMonth', 'gender_Male']  # AvgChargesPerMonth is almost identical to MonthlyCharges; gender_Male is duplicate
df_final = df_encoded.drop(columns=cols_to_drop, errors='ignore')

# Quick check
print("Columns dropped:", cols_to_drop)
df_final = df_final.rename(columns={"Churn_Yes": "Churn"})
df_final.head()


df = df.drop(columns=['ChurnFlag'], errors='ignore')

# Convert all boolean columns to int (0/1)
bool_cols = df_final.select_dtypes(include=['bool']).columns
df_final[bool_cols] = df_final[bool_cols].astype(int)

# Save the final preprocessed dataset
final_path = "/home/danial/Data Science/Churn Prediction/Data/Processed/final_preprocessed.csv"
df_final.to_csv(final_path, index=False)
print(f"Final dataset saved to: {final_path}")

# Quick check
df_final.head()
